Issue 763:	No mdRef to descriptive metadata in METS file for DSpace exports
Autoconverted from SVN (revision:1887)

+++ localDevSetup/cleanup.sh
@@ -1,6 +1,6 @@
-sudo rm /usr/share/fits/xml/fits.xml
-sudo mv /usr/share/fits/xml/fits.xml.backup /usr/share/fits/xml/fits.xml
-
+if [ -e /usr/share/fits/xml/fits.xmlbackup ]; then
+    sudo rm /usr/share/fits/xml/fits.xml
+fi
 
 sudo rm -r /usr/lib/archivematica
 sudo rm -r /etc/archivematica
